validate date in component, strings fix
