+++ demoapp/src/main/java/com/mercadolibre/android/andesui/demoapp/feature/DatePickerShowcaseActivity.kt
@@ -86,12 +86,12 @@ class DatePickerShowcaseActivity : AppCompatActivity() {
                 datepicker.clearMinMaxDate()
                 var setterMax: String? = inputMaxDate.text?.trim()
                 var setterMin: String? = inputMinDate.text?.trim()
-                if (setterMax != null && isValid(setterMax,"dd/MM/yyyy") && !setterMax.isEmpty()) {
+                if (setterMax != null && datepicker.isValid(setterMax,"dd/MM/yyyy") && !setterMax.isEmpty()) {
                     datepicker.setupMaxDate(setterMax, "dd/MM/yyyy" )
                 }else {
                     Toast.makeText(context, "la fecha maxima no es una fecha valida", Toast.LENGTH_SHORT).show()
                 }
-                if (setterMin != null && isValid(setterMin, "dd/MM/yyyy") && !setterMin.isEmpty()) {
+                if (setterMin != null && datepicker.isValid(setterMin, "dd/MM/yyyy") && !setterMin.isEmpty()) {
                     datepicker.setupMinDate(setterMin.toString(),"dd/MM/yyyy" )
                 }else {
                     Toast.makeText(context, "la fecha minima no es una fecha valida", Toast.LENGTH_SHORT).show()
@@ -111,17 +111,6 @@ class DatePickerShowcaseActivity : AppCompatActivity() {
 
             return layoutDatePicker
 
-        }
-
-        fun isValid(time:String, format:String): Boolean {
-            val df = SimpleDateFormat(format)
-            df.isLenient = false
-            try {
-                df.parse(time)
-                return true
-            } catch (e: ParseException){
-                return false
-            }
         }
 
     }
