+++ demoapp/src/main/res/values/strings.xml
@@ -220,7 +220,6 @@
     <string name="andesui_demoapp_bottom_sheet_set_button">Set</string>
     <string name="andesui_demoapp_bottom_sheet_dummy_view">This is a dummy view</string>
 
-
     <string name="andesui_demoapp_datepicker_text">Aplicar</string>
     <string name="andesui_demoapp_datepicker_button">Andes DatePicker</string>
     <string name="andesui_demoapp_datepicker_mindate_label">Setup min date</string>
@@ -228,12 +227,6 @@
     <string name="andesui_demoapp_datepicker_placeholder">dd/MM/yyyy</string>
     <string name="andesui_demoapp_datepicker_btn_send">Set MIN MAX Date</string>
     <string name="andesui_demoapp_datepicker_btn_reset">Reset</string>
-
-
-
-
-
-
 
     <string name="andesui_demoapp_loading_title">Click the button to show the spinner</string>
